refactor: code style (#1181)
* refactor: code style

* refactor: mv Language
