+++ src/components/Posts/PostBody.tsx
@@ -11,11 +11,11 @@ import { useAsync } from 'react-use';
 import Lock from '@/assets/lock.svg';
 import { Frame } from '@/components/Frame/index.js';
 import { NakedMarkup } from '@/components/Markup/NakedMarkup.js';
+import { PostMarkup } from '@/components/Markup/PostMarkup.js';
 import { Oembed } from '@/components/Oembed/index.js';
 import { Attachments } from '@/components/Posts/Attachment.js';
 import { CollapsedContent } from '@/components/Posts/CollapsedContent.js';
 import { ContentTranslator } from '@/components/Posts/ContentTranslator.js';
-import { PostMarkup } from '@/components/Posts/PostMarkup.js';
 import { Quote } from '@/components/Posts/Quote.js';
 import { IS_APPLE, IS_SAFARI } from '@/constants/bowser.js';
 import { STATUS } from '@/constants/enum.js';
@@ -70,9 +70,10 @@ export const PostBody = forwardRef<HTMLDivElement, PostBodyProps>(function PostB
 
     const muted = useIsMuted(post.author);
 
-    const postContent = (endingLinkCollapsed
-        ? removeUrlAtEnd(post.metadata.content?.oembedUrl, post.metadata.content?.content)
-        : post.metadata.content?.content) ?? '';
+    const postContent =
+        (endingLinkCollapsed
+            ? removeUrlAtEnd(post.metadata.content?.oembedUrl, post.metadata.content?.content)
+            : post.metadata.content?.content) ?? '';
 
     if (post.isEncrypted) {
         return (
